cold-start-predictor/requirements.txt
fastapi==0.135.3
pydantic==2.13.0
numpy==1.26.4
pyyaml==6.0.3
tensorflow-cpu==2.10.0
scikit-learn==1.7.2
uvicorn==0.44.0
python-multipart==0.0.9
requests==2.33.1
mangum==0.17.0
h5py==3.16.0
pickle5==0.0.12 ; python_version < "3.8"
pandas==2.3.3
scipy==1.15.3
pydantic_core==2.46.0
typing_extensions==4.15.0
anyio==4.13.0
starlette==1.0.0
boto3==1.34.0
python-dotenv==1.0.1
